--- Alunos/ARTUR ROBERTO MULLER KICH/AULA 01/src/Main.java
@@ -1,9 +1,0 @@
-//TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
-// click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("HELLO WORLD!");
-    }
-    }
-
-
