+++ Alunos/ARTUR ROBERTO MULLER KICH/Aula01/Hello.java
@@ -1,0 +1,8 @@
+
+public class Hello {
+    public static void main(String[] args) {
+        System.out.println("HELLO WORLD!");
+    }
+    }
+
+
